refactor: add database folder, prepare to extract the view
