+++ frontend/rust-lib/flowy-database/src/services/database/mod.rs
@@ -1,0 +1,10 @@
+mod block_editor;
+mod block_manager;
+mod database_editor;
+mod retry;
+mod trait_impl;
+
+pub use block_editor::*;
+pub use block_manager::*;
+pub use database_editor::*;
+pub use trait_impl::*;
