+++ frontend/rust-lib/flowy-database/src/services/database/trait_impl.rs
@@ -1,18 +1,19 @@
 use crate::entities::FieldType;
-use crate::services::block_manager::DatabaseBlockManager;
 use crate::services::cell::AtomicCellDataCache;
+use crate::services::database::DatabaseBlockManager;
+use crate::services::database_view::DatabaseViewEditorDelegate;
 use crate::services::field::{TypeOptionCellDataHandler, TypeOptionCellExt};
 use crate::services::row::DatabaseBlockRowRevision;
-use crate::services::view_editor::DatabaseViewEditorDelegate;
 
 use flowy_client_sync::client_database::DatabaseRevisionPad;
 use flowy_task::TaskDispatcher;
 use grid_model::{FieldRevision, RowRevision};
 use lib_infra::future::{to_fut, Fut};
+use std::any::type_name;
 use std::sync::Arc;
 use tokio::sync::RwLock;
 
-pub(crate) struct DatabaseViewEditorDelegateImpl {
+pub struct DatabaseViewEditorDelegateImpl {
   pub(crate) pad: Arc<RwLock<DatabaseRevisionPad>>,
   pub(crate) block_manager: Arc<DatabaseBlockManager>,
   pub(crate) task_scheduler: Arc<RwLock<TaskDispatcher>>,
@@ -27,7 +28,8 @@ impl DatabaseViewEditorDelegate for DatabaseViewEditorDelegateImpl {
         Ok(field_revs) => field_revs,
         Err(e) => {
           tracing::error!(
-            "[DatabaseViewEditorDelegate] get field revisions failed: {}",
+            "[{}] get field revisions failed: {}",
+            type_name::<DatabaseViewEditorDelegateImpl>(),
             e
           );
           vec![]
