+++ frontend/rust-lib/flowy-database/src/services/database_view/mod.rs
@@ -1,8 +1,8 @@
-mod changed_notifier;
 mod editor;
 mod editor_manager;
+mod notifier;
 mod trait_impl;
 
-pub use changed_notifier::*;
 pub use editor::*;
 pub use editor_manager::*;
+pub use notifier::*;
