+++ frontend/rust-lib/flowy-database/src/services/mod.rs
@@ -1,16 +1,12 @@
 mod util;
 
-pub mod block_editor;
-pub mod block_manager;
 pub mod cell;
-pub mod database_editor;
+pub mod database;
+pub mod database_view;
 pub mod field;
 pub mod filter;
 pub mod group;
 pub mod persistence;
-mod retry;
 pub mod row;
 pub mod setting;
 pub mod sort;
-mod trait_impl;
-pub mod view_editor;
