+++ frontend/rust-lib/flowy-database/src/services/sort/controller.rs
@@ -1,11 +1,11 @@
 use crate::entities::FieldType;
 use crate::entities::SortChangesetNotificationPB;
 use crate::services::cell::{AtomicCellDataCache, TypeCellData};
+use crate::services::database_view::{DatabaseViewChanged, DatabaseViewChangedNotifier};
 use crate::services::field::{default_order, TypeOptionCellExt};
 use crate::services::sort::{
   ReorderAllRowsResult, ReorderSingleRowResult, SortChangeset, SortType,
 };
-use crate::services::view_editor::{DatabaseViewChanged, GridViewChangedNotifier};
 use flowy_error::FlowyResult;
 use flowy_task::{QualityOfService, Task, TaskContent, TaskDispatcher};
 use grid_model::{CellRevision, FieldRevision, RowRevision, SortCondition, SortRevision};
@@ -34,7 +34,7 @@ pub struct SortController {
   sorts: Vec<Arc<SortRevision>>,
   cell_data_cache: AtomicCellDataCache,
   row_index_cache: HashMap<String, usize>,
-  notifier: GridViewChangedNotifier,
+  notifier: DatabaseViewChangedNotifier,
 }
 
 impl SortController {
@@ -45,7 +45,7 @@ impl SortController {
     delegate: T,
     task_scheduler: Arc<RwLock<TaskDispatcher>>,
     cell_data_cache: AtomicCellDataCache,
-    notifier: GridViewChangedNotifier,
+    notifier: DatabaseViewChangedNotifier,
   ) -> Self
   where
     T: SortDelegate + 'static,
